==> 02_3T3L1_prot/04_proteins_of_interest.Rmd <==
---
title: "04_proteins_of_interest"
output: html_document
---

###Background
Here I inspect protein groups of interest

###Initialise
```{r}
#Packages
library(ggplot2)

#Scripts
source("..\\scripts/essential_visualisations_JvG.R")
source("scripts/01_objects.R")
source("scripts/02_functions.R")

##Load in data
#Proteome
load("data/intermediate/proteome_post_analysis_3t3.RData")
```


###Phosphatases:
```{r}
###Load in
load("data/intermediate/phosphatase_data/pptase.gene.RData")
pptase_genes <- pptase.gene

##Process
pptase_genes <- sapply(pptase_genes,
                       function(x){
                         x <- tolower(x)
                         split_x <- strsplit(x, "")[[1]]
                         return(paste(c(toupper(split_x[1]),
                                        split_x[-1]),
                                      collapse = ""))
                       })
pptase_genes[which(pptase_genes == "Tenc1")] <- "Tns2"



up_geq1_ppases_in_proteome <- proteome_post_analysis_3t3[which(proteome_post_analysis_3t3$gene %in%
                                                                 pptase_genes &
                                                                 proteome_post_analysis_3t3$numisupmodels > 0), 
                                                         "gene"]
down_geq1_ppases_in_proteome <- proteome_post_analysis_3t3[which(proteome_post_analysis_3t3$gene %in%
                                                                 pptase_genes &
                                                                 proteome_post_analysis_3t3$numisdownmodels > 0), 
                                                         "gene"]

###Visualise

##Heatmap
  
fc_sig_heatmap(data = proteome_post_analysis_3t3[which(proteome_post_analysis_3t3$gene %in%
                                                                         pptase_genes &
                                                                         proteome_post_analysis_3t3$numisupordownmodels > 0), ],
                        fc_cols_w_pvals = c("CI_GROUPED_FC",
                                            "DEX_GROUPED_FC",
                                            "TNF_GROUPED_FC",
                                            "MPQ_GROUPED_FC",
                                            "AA_GROUPED_FC"),
                        pval_cols = c("CI_GROUPED_adj_pval",
                                            "DEX_GROUPED_adj_pval",
                                            "TNF_GROUPED_adj_pval",
                                            "MPQ_GROUPED_adj_pval",
                                            "AA_GROUPED_adj_pval"),
                        x_axis_names = c("CI",
                                         "DEX",
                                         "TNF",
                                         "MPQ",
                                         "AA"),
                        order_column = "CI_GROUPED_FC",
               yaxis_naming = "gene_uniprot") +
  ggtitle("Pptases changed in 1 or more models") 
  ggsave("output/images/analysis/proteins_of_interest/pptases_geq1.pdf",
         height = 6,
         width  = 6)

###All changed pptases, output data
write.csv(proteome_post_analysis_3t3[which(proteome_post_analysis_3t3$gene %in%
                                   pptase_genes &
                                   (proteome_post_analysis_3t3$numisupmodels > 0 |
                                      proteome_post_analysis_3t3$numisdownmodels > 0)), ],
          file = "output/data/3t3l1_proteome_altered_pptases.csv")
```
